[focus_areas] Simplify file names
File names for the description of focus areas included the
work "goal", which is no longer valid, and makes the names overly
long. So, I'm removing it.

I also removed them from the directories in which they were,
moving them directly to the focus_area directory. Right now
I'm not sure if we're going to need the directories again at some
point, but we can easily recreate them if needed.

I'm also simplifying the name of the directory for templates.

The main README.md directory of the repository already linked
to these file names.