Improve Code_Changes implementation and metric description
This is a more elaborate case of how a metric description,
and its reference implementation, can be.

I've included content to complete the description of the metric,
with the parameters needed to understand it, and the specific
stuff relevant for the case of git.

I've improved the notebook for implementing it.

The patch includes also some typos fixed.