Update Issue_Age.md
I made a couple of improvements in the text to make it more clear.
I also have some suggestions and comments that are NOT YET integrated in the text: (1) I do not tend to agree with the proposed "Implementation" description, that does not currently include counting all issues that have been opened AND closed in the considered time period. I would include these as well in the metric. In other words, I do not see a reason to restrict to only those issues that are still open at the end of the considered period. If you want to know the average age that an issue stays open, IMHO you should also include those issues that are already closed and count how long they have stayed open before they were closed.
(2) Always be careful with an "Average" aggregator. Average is only a good aggregator is data is normally distributed, but is misleading for skewed distributions. For such distributions, other aggregators, like for example "Median" will provide more robust results. This is a well-known fact in statistical analysis. This remark is likely to be applicable to any other CHAOSS metric that relies on the use of "Average" as an aggregator.
(3) How does the metric deal with incomplete data? For example, suppose that the issue was opened already before the start date of the considered period, then should I include the issue age as part of the metric? If yes, how? If we do not know when the issue was opened (since it happened before the considered period, also known as left-censoring) then we cannot correctly compute its age. One solution would be to exclude such issues from the aggregation.

Signed-off-by: Tom Mens <tom.mens.be@gmail.com>