Update New_Contributor_Closing_Issues.md (#297)
I provided some extensions of the metric description (e.g. a second aggregator and a second visualisation) and some minor improvements. Feel free to ignore any of these if you do not agree.
In addition, I have some comments/questions I did not integrate: (1) First of all, I think the metric name "New Contributors Closing Issues" is misleading, since the contributors are not necessarily "New". They may have been there for a long time before they actually closed an issue for the first time. (2) Secondly, this metric seems specific to GitHub Issues, but a very similar metric could also be defined for "Closing Pull Requests". So either the current metric could be generalised to also include closing of pull requests (could be a parameter of the metric), or another metric could be proposed that is basically the same, except that it works on pull requests rather than on issues. (3) Thirdly, I do not see why the filer "exclude reopened issues" needs to be present. Even if an issue has been reopened that was closed by a first-time-issue-closer, it seems strange to not include that specific contributor in the count, just because someone decides to reopen the issue later.